Override Kafka Connect distributed user's shell
A Puppet user resource uses the '/bin/sh' shell which allows login to this
service account. This commit allows to override this value and set
'/usr/sbin/nologin' for example. The default behavior is to left the shell
untouched in case the nologin shell is not available on the managed system.

Signed-off-by: Julien Riou <julien@riou.xyz>